[ai-assisted] refactor(layout): split full layout responsibilities
Why:
- Address issue #59 by reducing FullLayout responsibility and isolating navigation and user menu concerns.

What:
- Extracted navigation config/rendering into FullLayoutNavigation.
- Extracted header user/profile/theme menu into FullLayoutUserMenu.
- Kept FullLayout focused on shell state, header composition, drawer composition, and Outlet rendering.
- Updated CHANGELOG.md for issue #59.

Validation:
- npm run typecheck: pass
- npm run lint: pass with existing warnings only
- npm run build: pass; Vite reported existing large chunk warning
